Move Yaffs driver to a separate compilation unit
Extract the ioctl-based Yaffs driver code from src/mtd.c to src/ydrv.c,
so that the layers of abstraction involved are more accurately
represented by the organization of the source tree.  Prevent breaking
the build by temporarily retaining some duplicate code in both
compilation units.
