+++ src/ydrv.h
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: Michał Kępień <yafut@kempniu.pl>
+//
+// SPDX-License-Identifier: GPL-2.0-only
+
+#include <yaffs_guts.h>
+
+extern const struct yaffs_driver yaffs_driver_mtd;
